Desktop/ml_project/requirements.txt
# ── ML Studio — Python dependencies ──────────────────────────────────────────
# Tech stack: Python · Supabase · Streamlit  (no Docker, no Node.js)
# Install:  pip install -r requirements.txt
# Run:      streamlit run app.py

# UI
streamlit>=1.35.0
plotly>=5.20.0

# ML
scikit-learn>=1.4.0
pandas>=2.0.0
numpy>=1.26.0
joblib>=1.3.0

# File formats
openpyxl>=3.1.0
xlrd>=2.0.0
pyarrow>=14.0.0
odfpy>=1.4.1

# Supabase
supabase>=2.4.0
python-dotenv>=1.0.0
